Fix websocket not passing proper post details
When a new post was posted, only its body was sent accross as a websocket message causing the clients (incuding the senders browser) to not be able to display the post authors details as those were populated using a foreign key and mongoose's populate function when the posts were extracted from the DB.

This meant that after refreshing the page (all posts are fetched from DB not websocket) the details were there but not when a post was recieved using websocket.

To fix this instead of taking the post object, saving it then sending the initial post object accross the websockets, after saving the post object, I pull it out of the DB and send that. This means that the post object is populated with all the required author details. (Sending it the previous way mean that all the post object had was the foreign key, which the other client can't do much with)