Prevent animation on initial render of page
This way on the page load not all the elements are animated as that look ridiculus. Wrapping the the list of posts inside an ng-if="posts.lenght" means that the list is hidden until it is poulated with something. This means that the initial animation to populat the page will not be seen by the user